動く

+++ Dataset.py
@@ -1,6 +1,7 @@
 import torch
 import numpy as np
 
+
 class MyDataset(torch.utils.data.Dataset):
     def __init__(self, data_num, transform=None):
         self.transform = transform
@@ -8,8 +9,8 @@ class MyDataset(torch.utils.data.Dataset):
         self.data = []
         self.label = []
         for x in range(self.data_num):
-            self.data.append(np.zeros([24, 256, 256]))  # 0 から (data_num-1) までのリスト
-            self.label.append(x % 2 == 0)  # 偶数ならTrue 奇数ならFalse
+            self.data.append(np.zeros([24, 256, 256], dtype=np.float32))  # 0 から (data_num-1) までのリスト
+            self.label.append(np.float32(x % 2 == 0))  # 偶数ならTrue 奇数ならFalse
 
     def __len__(self):
         return self.data_num
